refactor : validate 책임을 변경 및 오버로딩 활용으로 다형성 구현
