+++ src/main/java/lotto/domain/common/EntryValidator.java
@@ -1,8 +1,11 @@
 package lotto.domain.common;
 
+import static lotto.domain.common.LottoConstant.LOTTO_COUNT_NUMBER;
 import static lotto.domain.common.LottoConstant.LOTTO_END_NUMBER;
 import static lotto.domain.common.LottoConstant.LOTTO_START_NUMBER;
+import static lotto.domain.common.error.ErrorCode.INVALID_LOTTO_SIZE;
 
+import java.util.Arrays;
 import lotto.domain.common.error.ErrorCode;
 import lotto.domain.common.error.exception.LottoIllegalArgumentException;
 
@@ -15,6 +18,11 @@ public class EntryValidator {
         validateEmpty(entry);
         validateNumber(entry);
         validateLottoRange(entry);
+    }
+
+    public static void validate(String[] numberArr){
+        validateLottoDistinct(numberArr);
+        validateEntrySize(numberArr);
     }
 
     private static void validateLottoRange(String i){
@@ -37,5 +45,22 @@ public class EntryValidator {
         if(entry.isEmpty()){
             throw new LottoIllegalArgumentException(ErrorCode. INVALID_EMPTY);
         }
+    }
+
+    private static void validateLottoDistinct(String[] numberArr){
+        String[] mapArr = distinctArr(numberArr);
+        if(mapArr.length!=LOTTO_COUNT_NUMBER){
+            throw new LottoIllegalArgumentException(INVALID_LOTTO_SIZE);
+        }
+    }
+
+    private static void validateEntrySize(String[] numberArr) {
+        if(numberArr.length!=LOTTO_COUNT_NUMBER){
+            throw new LottoIllegalArgumentException(INVALID_LOTTO_SIZE);
+        }
+    }
+
+    private static String[] distinctArr(String[] strArr){
+        return Arrays.stream(strArr).distinct().toArray(String[]::new);
     }
 }
